+++ lib/shopify_api/auth_token_server.ex
@@ -2,7 +2,6 @@ defmodule ShopifyAPI.AuthTokenServer do
   use GenServer
   require Logger
   alias ShopifyAPI.AuthToken
-  alias ShopifyAPI.EventPipe.EventQueue
 
   @name :shopify_api_auth_token_server
 
@@ -47,7 +46,6 @@ defmodule ShopifyAPI.AuthTokenServer do
 
   def set(%AuthToken{shop_name: shop, app_name: app} = token, false) do
     GenServer.cast(@name, {:set, AuthToken.create_key(shop, app), token})
-    Task.start(fn -> EventQueue.subscribe(token) end)
   end
 
   def set(%AuthToken{shop_name: shop, app_name: app} = token, true) do